requirements.txt
pymupdf
pypdf
pillow
